MCP-1983: separated mas integration routes for readability
