+++ service/provider/src/main/java/gov/va/vro/service/provider/CamelEntrance.java
@@ -2,6 +2,7 @@ package gov.va.vro.service.provider;
 
 import gov.va.vro.model.mas.MasAutomatedClaimPayload;
 import gov.va.vro.model.mas.MasExamOrderStatusPayload;
+import gov.va.vro.service.provider.camel.MasIntegrationRoutes;
 import gov.va.vro.service.provider.camel.PrimaryRoutes;
 import gov.va.vro.service.spi.model.Claim;
 import gov.va.vro.service.spi.model.GeneratePdfPayload;
@@ -42,11 +43,14 @@ public class CamelEntrance {
 
   public void notifyAutomatedClaim(MasAutomatedClaimPayload payload, long delay) {
     producerTemplate.sendBodyAndHeader(
-        PrimaryRoutes.ENDPOINT_AUTOMATED_CLAIM, payload, PrimaryRoutes.MAS_DELAY_PARAM, delay);
+        MasIntegrationRoutes.ENDPOINT_AUTOMATED_CLAIM,
+        payload,
+        MasIntegrationRoutes.MAS_DELAY_PARAM,
+        delay);
   }
 
   public String examOrderingStatus(MasExamOrderStatusPayload payload) {
-    producerTemplate.requestBody(PrimaryRoutes.ENDPOINT_EXAM_ORDER_STATUS, payload);
+    producerTemplate.requestBody(MasIntegrationRoutes.ENDPOINT_EXAM_ORDER_STATUS, payload);
     return "Message Received";
   }
 }
