+++ service/provider/src/main/java/gov/va/vro/service/provider/camel/PrimaryRoutes.java
@@ -1,18 +1,12 @@
 package gov.va.vro.service.provider.camel;
 
 import gov.va.vro.camel.FunctionProcessor;
-import gov.va.vro.model.event.JsonConverter;
-import gov.va.vro.model.mas.MasAutomatedClaimPayload;
 import gov.va.vro.service.event.AuditEventProcessor;
-import gov.va.vro.service.provider.MasPollingProcessor;
 import gov.va.vro.service.provider.services.AssessmentResultProcessor;
 import gov.va.vro.service.spi.db.SaveToDbService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.camel.Exchange;
-import org.apache.camel.ExchangePattern;
 import org.apache.camel.builder.RouteBuilder;
-import org.apache.camel.component.jackson.JacksonDataFormat;
 import org.springframework.stereotype.Component;
 
 /** Defines primary routes. */
@@ -21,19 +15,10 @@ import org.springframework.stereotype.Component;
 @RequiredArgsConstructor
 public class PrimaryRoutes extends RouteBuilder {
 
-  public static final String ENDPOINT_MAS =
-      "rabbitmq:mas-notification-exchange?queue=mas-notification-queue&routingKey=mas-notification&requestTimeout=0";
-
   public static final String ENDPOINT_SUBMIT_CLAIM = "direct:claim-submit";
   public static final String ENDPOINT_SUBMIT_CLAIM_FULL = "direct:claim-submit-full";
   public static final String ENDPOINT_GENERATE_PDF = "direct:generate-pdf";
   public static final String ENDPOINT_FETCH_PDF = "direct:fetch-pdf";
-
-  public static final String ENDPOINT_AUTOMATED_CLAIM = "direct:automated-claim";
-
-  public static final String ENDPOINT_EXAM_ORDER_STATUS = "direct:exam-order-status";
-
-  public static final String MAS_DELAY_PARAM = "masDelay";
 
   private static final String PDF_EXCHANGE = "pdf-generator";
   private static final String GENERATE_PDF_QUEUE = "generate-pdf";
@@ -41,21 +26,15 @@ public class PrimaryRoutes extends RouteBuilder {
 
   private final SaveToDbService saveToDbService;
 
-  private final SlipClaimSubmitRouter claimSubmitRouter;
-  private final MasPollingProcessor masPollingProcessor;
-
   private final AssessmentResultProcessor assessmentResultProcessor;
   private final AuditEventProcessor auditEventProcessor;
 
   @Override
   public void configure() {
-    configureExceptionHandling();
     configureRouteClaimSubmit();
     configureRouteClaimSubmitForFull();
     configureRouteGeneratePdf();
     configureRouteFetchPdf();
-    configureAutomatedClaim();
-    configureOrderExamStatus();
   }
 
   private void configureRouteClaimSubmit() {
@@ -91,50 +70,7 @@ public class PrimaryRoutes extends RouteBuilder {
     from(ENDPOINT_FETCH_PDF).routeId("fetch-pdf").to(pdfRoute(FETCH_PDF_QUEUE));
   }
 
-  private void configureAutomatedClaim() {
-    from(ENDPOINT_AUTOMATED_CLAIM)
-        .routeId("mas-claim-notification")
-        .process(
-            auditEventProcessor.event(
-                "mas-claim-notification",
-                "Setting a delay before staring Automated claim processing."))
-        .delay(header(MAS_DELAY_PARAM))
-        .setExchangePattern(ExchangePattern.InOnly)
-        .process(
-            auditEventProcessor.event("mas-claim-notification", "Calling endpoint " + ENDPOINT_MAS))
-        .to(ENDPOINT_MAS);
-
-    from(ENDPOINT_MAS)
-        .routeId("mas-claim-processing")
-        .unmarshal(new JacksonDataFormat(MasAutomatedClaimPayload.class))
-        .process(
-            auditEventProcessor.event("mas-claim-processing", "Entering endpoint " + ENDPOINT_MAS))
-        .process(masPollingProcessor)
-        .setExchangePattern(ExchangePattern.InOnly)
-        .log("MAS response: ${body}");
-  }
-
-  private void configureOrderExamStatus() {
-    String routeId = "exam-order-status";
-    from(ENDPOINT_EXAM_ORDER_STATUS)
-        .routeId(routeId)
-        .process(
-            auditEventProcessor.event(
-                routeId, "Entering endpoint " + ENDPOINT_EXAM_ORDER_STATUS, new JsonConverter()));
-  }
-
   private String pdfRoute(String queueName) {
     return String.format("rabbitmq:%s?routingKey=%s&queue=%s", PDF_EXCHANGE, queueName, queueName);
-  }
-
-  private void configureExceptionHandling() {
-    onException(Throwable.class)
-        .process(
-            exchange -> {
-              Throwable exception = (Throwable) exchange.getProperty(Exchange.EXCEPTION_CAUGHT);
-              var message = exchange.getMessage();
-              var body = message.getBody();
-              auditEventProcessor.logException(body, exception, exchange.getFromRouteId());
-            });
   }
 }
